Refactor into package structure
