+++ pyproject.toml
@@ -1,0 +1,17 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ghostlink"
+version = "0.1.0"
+description = "Stealth audio encoder for embedding structured text in music"
+readme = "README.md"
+authors = [{name = "GhostLink Contributors"}]
+license = {text = "MIT"}
+requires-python = ">=3.8"
+dependencies = []
+
+[project.scripts]
+ghostlink = "ghostlink.__main__:main"
+ghostlink-decode = "ghostlink.decoder:main"
