+++ tests/test_bits_to_symbols.py
@@ -1,8 +1,4 @@
-import os
-import sys
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from GhostLink import bits_to_symbols
+from ghostlink import bits_to_symbols
 
 
 def test_bits_to_symbols_preserves_input_list():
