+++ tests/test_hamming.py
@@ -1,9 +1,4 @@
-import os
-import sys
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-
-from GhostLink import hamming74_encode_nibble, hamming74_encode_bytes
+from ghostlink import hamming74_encode_nibble, hamming74_encode_bytes
 
 
 def test_hamming74_encode_nibble():
